+++ changelog.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Fixed
 
--
+- 'Multiple Characters' trimming within tables (#3)
 
 ---