+++ changelog.md
@@ -10,6 +10,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
+-
+
+### Changed
+
+-
+
+### Fixed
+
+-
+
+---
+
+## [v0.2.0] - 2022/07/13
+
+### Added
+
 - Setting to control minimum time delay between auto-trim (#1)
 - Option to skip code blocks (#4)
 
@@ -31,4 +47,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Initial commit!
 
 [Unreleased]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/master...develop
+[v0.2.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/a53bdb3...v0.1.0